Add (for now) read-only context-editor
It took quite some time to make the matrix look acceptable.
The next step is to be able to modify the context through
the matrix and the buttons. As far as I looked through the
source code there are no functions provided for the button
actions so I'd need to implement these, too.