Update weblaf jar
The open file dialog still doesn't work, but it seems to be a problem on our
side. I guess we need to adapt our code a bit. I put the weblaf demo jar in our
Dropbox folder so that we can take a look how to initialise the JFileChooser
correctly.